Import / export common functions
