+++ portfolio/src/main/webapp/index.html
@@ -10,9 +10,9 @@
       My Portfolio
     </title>
     <link rel="stylesheet" href="style.css">
-    <script src="script.js">
+    <script type="module" src="script.js">
     </script>
-    <script src="comments-script.js">
+    <script type="module" src="comments-script.js">
     </script>
     <script src="https://kit.fontawesome.com/150c4874c5.js" crossorigin="anonymous">
     </script>
